backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
httpx==0.28.1
PyMuPDF==1.26.3
pydantic==2.11.7
python-dotenv==1.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
